@@ -5,17 +5,18 @@
   "scripts": {
     "build:initial-project": "node ./tools/loadInitialProject.js",
     "build": "yarn run build:initial-project && webpack --colors",
+    "benchmark": "node ./benchmark/index.js",
     "dev": "yarn run build:initial-project && webpack-dev-server --watch --colors --inline --hot --config webpack.config.dev.js"
   },
   "dependencies": {
     "babel-polyfill": "^6.16.0",
     "prop-types": "^15.5.10",
     "ramda": "^0.24.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^16.2",
+    "react-dom": "^16.2",
     "react-event-listener": "^0.4.5",
     "react-fa": "^5.0.0",
-    "react-hotkeys": "git+https://github.com/xodio/react-hotkeys.git#npm",
+    "react-hotkeys": "^0.10.0",
     "react-redux": "^4.4.5",
     "react-skylight": "git+https://github.com/xodio/react-skylight.git",
     "redux": "^3.0.5",
@@ -26,6 +27,9 @@
     "xod-project": "^0.16.0"
   },
   "devDependencies": {
+    "node-static": "^0.7.10",
+    "puppeteer": "^0.13.0",
+    "why-did-you-update": "^0.1.0",
     "xod-fs": "^0.16.0"
   },
   "author": "XOD Team <dev@xod.io>",
