@@ -27,11 +27,11 @@
     "prop-types": "^15.5.10",
     "ramda": "^0.24.1",
     "rc-progress": "^2.1.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^16.2",
+    "react-dom": "^16.2",
     "react-event-listener": "^0.4.5",
     "react-fa": "^5.0.0",
-    "react-hotkeys": "git+https://github.com/xodio/react-hotkeys.git#npm",
+    "react-hotkeys": "^0.10.0",
     "react-redux": "^4.4.5",
     "react-skylight": "git+https://github.com/xodio/react-skylight.git",
     "redux": "^3.0.5",
@@ -53,6 +53,7 @@
     "devtron": "^1.3.0",
     "electron": "1.6.11",
     "electron-builder": "git+https://github.com/xodio/electron-builder#monorepo-fix-npm",
+    "electron-devtools-installer": "^2.2.1",
     "electron-mocha": "^3.4.0",
     "fsp": "^0.1.2",
     "webpack": "^1.13.2",
