@@ -21,7 +21,6 @@
     "classnames": "^2.2.5",
     "codemirror": "^5.31.0",
     "commander": "^2.9.0",
-    "debounce": "^1.1.0",
     "escape-string-regexp": "^1.0.5",
     "font-awesome": "^4.6.3",
     "formsy-react": "^0.18.1",
@@ -34,24 +33,24 @@
     "rc-menu": "^5.0.10",
     "rc-progress": "^2.1.2",
     "rc-switch": "^1.4.2",
-    "react": "^15.4.2",
+    "react": "^16.2",
     "react-autolink-text2": "git+https://github.com/xodio/react-autolink-text.git#npm",
     "react-autosuggest": "git+https://github.com/xodio/react-autosuggest.git#bc6ab71a2ec47ad5fbce5d13ff6a572942ea765e",
     "react-click-outside": "^2.2.0",
     "react-codemirror": "^1.0.0",
-    "react-collapsible": "git+https://github.com/xodio/react-collapsible.git#npm",
+    "react-collapsible": "^2.0.3",
     "react-custom-scroll": "^2.0.0",
     "react-dnd": "^2.5.1",
     "react-dnd-html5-backend": "^2.5.1",
-    "react-dom": "^15.4.2",
+    "react-dom": "^16.2",
     "react-event-listener": "^0.4.5",
     "react-fa": "^5.0.0",
     "react-highlight-words": "^0.8.1",
-    "react-hotkeys": "git+https://github.com/xodio/react-hotkeys.git#npm",
+    "react-hotkeys": "^0.10.0",
     "react-infinite": "^0.12.1",
     "react-redux": "^4.0.6",
     "react-skylight": "git+https://github.com/xodio/react-skylight.git",
-    "react-sortable-hoc": "0.0.7",
+    "react-sortable-hoc": "0.6.7",
     "react-split-pane": "^0.1.68",
     "recompose": "^0.25.0",
     "redux": "^3.0.5",
@@ -59,6 +58,7 @@
     "reselect": "^2.5.4",
     "sanctuary-def": "^0.9.0",
     "shortid": "^2.2.8",
+    "throttle-debounce": "^1.0.1",
     "url-parse": "^1.1.9",
     "url-search-params-polyfill": "^2.0.1",
     "xod-arduino": "^0.16.0",
@@ -71,10 +71,6 @@
     "@kadira/storybook": "^2.35.3",
     "chai": "^3.5.0",
     "dirty-chai": "^1.2.2",
-    "redux-devtools": "^3.4.0",
-    "redux-devtools-dispatch": "^2.2.1",
-    "redux-devtools-dock-monitor": "^1.1.2",
-    "redux-devtools-filterable-log-monitor": "^0.6.8",
     "redux-mock-store": "^1.2.3"
   }
 }
