@@ -58,9 +58,9 @@ module.exports = {
         '**/*.spec.js',
         '**/xod-client-electron/**/*.js',
         '**/xod-client-electron/**/*.jsx',
-        '**/xod-client/**/DevTools.jsx',
         '**/xod-client/stories/*.jsx',
-        '**/xod-client-browser/tools/*.js'
+        '**/xod-client-browser/tools/*.js',
+        '**/xod-client-browser/benchmark/*.js'
       ]
     }],
     'mocha/no-skipped-tests': 'error',
