@@ -2,7 +2,7 @@ import { compose, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 
 import resolveLibsMiddleware from '../project/resolveLibsMiddleware';
-import { devToolMiddleware } from '../utils/devtools';
+import devtoolsMiddleware from '../utils/devtoolsMiddleware';
 
 export default (extraMiddlewares = []) => compose(
   applyMiddleware(
@@ -10,5 +10,5 @@ export default (extraMiddlewares = []) => compose(
     resolveLibsMiddleware,
     ...extraMiddlewares
   ),
-  devToolMiddleware
+  devtoolsMiddleware
 );
